+++ exch/authmgr.hpp
@@ -3,17 +3,17 @@
 
 /**
  * %EXCH:	can login via emsmdb or zcore
- * %IMAP:	can login via IMAP
- * %PAM:	can login via pam_gromox with service==nullptr
+ * %IMAP:	can login via IMAP & POP3
  * %CHGPASSWD:	user is allowed to change their own password via zcore
  * %PUBADDR:	(unused)
- * %CHAT, %VIDEO, %FILES, %ARCHIVE:	pam_gromox with service=...
+ * %SMTP, %CHAT, %VIDEO, %FILES, %ARCHIVE:
+ * 		pam_gromox with service=...
  */
 enum {
 	USER_PRIVILEGE_EXCH = 0,
 	USER_PRIVILEGE_IMAP = 1U << 0,
 	USER_PRIVILEGE_POP3 = USER_PRIVILEGE_IMAP,
-	USER_PRIVILEGE_PAM = 1U << 1,
+	USER_PRIVILEGE_SMTP = 1U << 1,
 	USER_PRIVILEGE_CHGPASSWD = 1U << 2,
 	USER_PRIVILEGE_PUBADDR = 1U << 3,
 	USER_PRIVILEGE_CHAT = 1U << 4,